Do not duplicate groups

+++ build.fsx
@@ -102,30 +102,42 @@ type ServerPaketDependencies =
             let azure = if x.Azure then "azure" else "noazure"
             sprintf "%s-%s-%s" server remoting azure
 
+type CombinedPaketDependencies =
+    { Azure : bool
+      Remoting : bool
+      Fulma : bool
+      Server : ServerPaketDependency }
+
+    member x.ToBuild : BuildPaketDependencies =
+            { Azure = x.Azure }
+
+    member x.ToClient : ClientPaketDependencies =
+        { Remoting = x.Remoting
+          Fulma = x.Fulma }
+
+    member x.ToServer : ServerPaketDependencies =
+        { Server = x.Server
+          Remoting = x.Remoting
+          Azure = x.Azure }
+
 let configs =
     [ "",
-      ( { Azure = false },
-        { Remoting = false
-          Fulma = true },
-        { Server = Saturn
-          Remoting = false
-          Azure = false } )
+      { Azure = false
+        Fulma = true
+        Server = Saturn
+        Remoting = false }
 
       "--deploy azure",
-      ( { Azure = true },
-        { Remoting = false
-          Fulma = true },
-        { Server = Saturn
-          Remoting = false
-          Azure = true } )
+      { Azure = true
+        Fulma = true
+        Server = Saturn
+        Remoting = false }
 
       "--layout none",
-      ( { Azure = false },
-        { Remoting = false
-          Fulma = false },
-        { Server = Saturn
-          Remoting = false
-          Azure = false } )
+      { Azure = false
+        Fulma = false
+        Server = Saturn
+        Remoting = false }
     ]
 
 
@@ -133,27 +145,27 @@ let fullLockFileName build client server =
     sprintf "paket_%O_%O_%O.lock" build client server
 
 Target.create "BuildPaketLockFiles" (fun _ ->
-    for (_, (build, client, server)) in configs do
+    for (_, config) in configs do
         let contents =
             [
-                "Content" </> "src" </> "Build" </> sprintf "paket_%O.lock" build
-                "Content" </> "src" </> "Client" </> sprintf "paket_%O.lock" client
-                "Content" </> "src" </> "Server" </> sprintf "paket_%O.lock" server
+                "Content" </> "src" </> "Build" </> sprintf "paket_%O.lock" config.ToBuild
+                "Content" </> "src" </> "Client" </> sprintf "paket_%O.lock" config.ToClient
+                "Content" </> "src" </> "Server" </> sprintf "paket_%O.lock" config.ToServer
             ]
             |> List.map File.read
             |> Seq.concat
 
-        let lockFileName = fullLockFileName build client server
+        let lockFileName = fullLockFileName config.ToBuild config.ToClient config.ToServer
 
         File.writeWithEncoding (Text.UTF8Encoding(false)) false ("Content" </> lockFileName) contents
 )
 
 Target.create "GenJsonConditions" (fun _ ->
-    for (_, (build, client, server)) in configs do //TODO this combination is different?
-        let lockFileName = fullLockFileName build client server
+    for (_, config) in configs do //TODO this combination is different?
+        let lockFileName = fullLockFileName config.ToBuild config.ToClient config.ToServer
         let server = "saturn"
-        let deploy = if build.Azure then "azure" else "none"
-        let layoutOperator = if client.Fulma then "!=" else "=="
+        let deploy = if config.Azure then "azure" else "none"
+        let layoutOperator = if config.Fulma then "!=" else "=="
         let template =
             sprintf """                    {
                         "include": "%s",
@@ -171,11 +183,15 @@ Target.create "GenJsonConditions" (fun _ ->
 )
 
 Target.create "GenPaketLockFiles" (fun _ ->
-    let dirName = "gen-paket-lock-files"
-    Directory.delete dirName
-    Directory.create dirName
+    let baseDir = "gen-paket-lock-files"
+    Directory.delete baseDir
+    Directory.create baseDir
 
-    for (arg, (build, client, server)) in configs do
+    for (arg, config) in configs do
+        let dirName = baseDir </> "tmp"
+        Directory.delete dirName
+        Directory.create dirName
+
         run "dotnet" (sprintf "new SAFE %s" arg) dirName
 
         let lockFile = dirName </> "paket.lock"
@@ -194,17 +210,19 @@ Target.create "GenPaketLockFiles" (fun _ ->
             |> List.filter (String.isNullOrWhiteSpace >> not)
             |> List.map (fun group -> group.Substring(0, group.IndexOf Environment.NewLine), delimeter + group)
         for (groupName, group) in groups do
-            let dirName = dirName </> groupName
+            let dirName = baseDir </> groupName
             Directory.create dirName
             let lockFileSuffix =
                 match groupName with
-                | "Build" -> string build
-                | "Client" -> string client
-                | "Server" -> string server
+                | "Build" -> string config.ToBuild
+                | "Client" -> string config.ToClient
+                | "Server" -> string config.ToServer
                 | _ -> failwithf "Unhandled name '%s'" groupName
             let fileName = sprintf "paket_%s.lock" lockFileSuffix
-            File.writeString false (dirName </> fileName) group
-            Shell.copyFile ("Content" </> "src" </> groupName </> fileName) (dirName </> fileName)
+            let filePath = dirName </> fileName
+            if not (File.exists filePath) then
+                File.writeString false filePath group
+                Shell.copyFile ("Content" </> "src" </> groupName </> fileName) (dirName </> fileName)
 )
 
 Target.create "Tests" (fun _ ->
@@ -248,5 +266,8 @@ open Fake.Core.TargetOperators
     ==> "Tests"
     ==> "Push"
     ==> "Release"
+
+"Install"
+    ==> "GenPaketLockFiles"
 
 Target.runOrDefault "Pack"
